Moved ISO2 Mapping from a phone number out of E164PhoneNumber class into an own class ISO2Mapper class, added to that one, ISO2 Mapping from country name used in DT online price list.
Also Added Data consistency test for both mappings and discovered, that 59 countries are not listed in the official pricelist. Tested a few of them by trying to send an SMS by the API via Postman and got a new error message, that routing to those numbers is not supported. So introduced a new list for those country codes, to have consistent groups of ISO2 code categories.